add status validation
