+++ .pre-commit-config.yaml
@@ -1,8 +1,17 @@
 repos:
   - repo: local
     hooks:
+      - id: adr-status-valid
+        name: Validate ADR status
+        entry: hack/adr-tool.py
+        args:
+          - validate
+        language: python
+        files: ^docs/modules/ROOT/pages/adr/
       - id: adr-index
         name: Generate ADR Index
-        entry: hack/generate-adr-index.py
+        entry: hack/adr-tool.py
+        args:
+          - generate
         language: python
         files: ^docs/modules/ROOT/pages/adr/
